+++ app/(test)/taxonomy/search/loading.tsx
